+++ src/strategies/php-yoshi.ts
@@ -194,6 +194,29 @@ export class PHPYoshi extends BaseStrategy {
     };
   }
 
+  protected async parsePullRequestBody(
+    pullRequestBody: string
+  ): Promise<PullRequestBody | undefined> {
+    const body = PullRequestBody.parse(pullRequestBody);
+    if (!body) {
+      return undefined;
+    }
+    const component = await this.getComponent();
+    const notes = body.releaseData
+      .map(release => {
+        return `<details><summary>${
+          release.component
+        }: ${release.version?.toString()}</summary>\n\n${
+          release.notes
+        }\n</details>`;
+      })
+      .join('\n\n');
+    return new PullRequestBody([{component, notes}], {
+      footer: body.footer,
+      header: body.header,
+    });
+  }
+
   protected async buildUpdates(
     options: BuildUpdatesOptions
   ): Promise<Update[]> {