+++ test/updaters/composite.ts
@@ -12,11 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {CompositeUpdater} from '../../src/updaters/composite';
+import {CompositeUpdater, mergeUpdates} from '../../src/updaters/composite';
 import {JavaUpdate} from '../../src/updaters/java/java-update';
 import {Version} from '../../src/version';
 import {describe, it} from 'mocha';
 import {expect} from 'chai';
+import {Updater} from '../../src/update';
+import {assertHasUpdate} from '../helpers';
+
+class FakeUpdater implements Updater {
+  updateContent(_content: string | undefined): string {
+    return '';
+  }
+}
 
 describe('CompositeUpdater', () => {
   describe('updateContent', () => {
@@ -42,6 +50,66 @@ describe('CompositeUpdater', () => {
       expect(output).to.eql(
         'v1: 1.2.3 // {x-version-update:artifact1:current}\nv2: 2.0.0 // {x-version-update:artifact2:current}'
       );
+    });
+  });
+  describe('mergeUpdates', () => {
+    it('can merge multiple updates', () => {
+      const input = [
+        {
+          path: 'path1',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path2',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path1',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path1',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+      ];
+      const merged = mergeUpdates(input);
+      expect(merged).lengthOf(2);
+      assertHasUpdate(merged, 'path1', CompositeUpdater);
+      assertHasUpdate(merged, 'path2', FakeUpdater);
+    });
+    it('ignores disjoint paths', () => {
+      const input = [
+        {
+          path: 'path1',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path2',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path3',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+        {
+          path: 'path4',
+          createIfMissing: false,
+          updater: new FakeUpdater(),
+        },
+      ];
+      const merged = mergeUpdates(input);
+      expect(merged).lengthOf(4);
+      assertHasUpdate(merged, 'path1', FakeUpdater);
+      assertHasUpdate(merged, 'path2', FakeUpdater);
+      assertHasUpdate(merged, 'path3', FakeUpdater);
+      assertHasUpdate(merged, 'path4', FakeUpdater);
     });
   });
 });